pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kdv"
version = "0.1.0"
description = "Kdenlive Video Workflow Toolkit for drone footage"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Jay Firns"}
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
kdv = "kdv.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["kdv*"]
